requirements.txt
requests==2.31.0
mysql-connector-python==8.2.0
APScheduler==3.10.4
python-dotenv==1.0.0
